Specify base spec of the gomoku game
Players were not intitally thought of when designing the interface. They
could be introduced later, and just pretend the information was unknown
at the point. The cost of doing so was modifying the core path.

On second thought, the core path would not be complete without the idea
of players. MVP should not mean being spearheaded and only focusing on one
tiniest single important thing at a time.